Cluster CIDR check fix

+++ calico-cluster-check.sh
@@ -111,11 +111,13 @@ function check_cluster_pod_cidr {
                 pod_cidr=`kubectl get ippool -o yaml | grep cidr | grep -o '[0-9]\{1,3\}\.[0-9]\{1,3\}\.[0-9]\{1,3\}\.[0-9]\{1,3\}\/[1-9]\{1,2\}'`
 		if [ $pod_cidr == '' ]; then echo "Unable to retrieve the pod cidr information"; else echo "The pod cidr is $pod_cidr"; fi
 		
-		cidr_check=$(cidr_check_status $cluster_cidr $pod_cidr)
+		if [  ! -z $pod_cidr ]  &&  [ ! -z $cluster_cidr ]; then cidr_check=$(cidr_check_status $cluster_cidr $pod_cidr); fi
 
 		if [ "$cidr_check" == "True" ] && [ ! -z "$cluster_cidr" ]; then echo "Pod cidr is a subset of Cluster cidr"; else echo ""; fi
 
-		rm cidrcheck.py
+		if [ -f cidrcheck.py ]; then
+			rm cidrcheck.py
+		fi
 
                 echo -e "\n"
 }
